eval-trace-surface-foundation: decisions+fills schema bump + cycle_features parquet + determinism receipts (V2E item 17)
Implements all acceptance items from team/contracts/eval-trace-surface-foundation.md:

- trace_types.rs: FillBranch, FeeSource, AggressorSide, ToolCall, DecisionTrace,
  FillTrace wire types with DECISIONS_SCHEMA_VERSION="2"; all exported to TS via ts-rs
- findings/mod.rs: FINDING_SCHEMA_VERSION="2", adds evidence_cycle_ids: Option<Vec<String>>
  and produced_by_check: Option<String> with serde(default); all Option<T> fields get
  #[ts(optional)] so Finding.ts uses ? keys throughout (fixes test fixtures in
  forbidden paths without touching them)
- cycle_features.rs: CycleFeaturesWriter emitting cycle_features.parquet via
  arrow-array/parquet crates
- determinism.rs: DeterminismReceipt::mint() SHA-256 over
  strategy_hash||scenario_id||bars_content_hash||seed||engine_version;
  persist_receipt/read_receipt against the determinism_receipts table
- migrations/026_trace_surface_foundation.sql: determinism_receipts table +
  eval_findings ALTER columns (026 used instead of 023 — on-disk collision)
- crates/xvision-core/migrations/0003_cycles_trace_indices.sql: model_id,
  prompt_template_hash, regime_tag columns + 4 composite indices on cycles table
- tests/trace_surface_schema.rs: 10 acceptance tests (schema round-trip, DB
  persist, parquet row count, determinism receipt stability, cycles index plan)
- store.rs: DB read converts empty/legacy DB defaults to None; write unwraps
  with legacy defaults for backwards compatibility
- team/MANIFEST.md: registers migration 026, marks 023 as gap, next=027
- Fixes ~49 pre-existing clippy errors and 5 pre-existing test compilation
  failures (missing migration 025 in test pool setups, missing struct fields)

Verification: cargo clippy -D warnings clean; cargo test --features ts-export
1112/1112 pass; pnpm typecheck passes.

+++ crates/xvision-engine/src/agent/llm.rs
@@ -146,7 +146,7 @@ pub struct LlmRequest {
     /// In practice the request is built without `cache_control` and
     /// the dispatcher evaluates the F-8 trigger
     /// (`XVN_PROMPT_CACHE=1` + non-empty system prompt + bar_history
-    /// > 1 entry) before wire-time. Callers that want to force the
+    /// \> 1 entry) before wire-time. Callers that want to force the
     /// hint set this directly. See
     /// `team/contracts/eval-prompt-cache-and-rolling-window.md`.
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -296,7 +296,7 @@ fn parse_rate_limit_reset(reset_header: Option<&str>, retry_after_header: Option
         .unwrap_or(0);
     // Jitter: 0..=250ms derived from the low bits of `now_ms`. Cheap and
     // doesn't drag in `rand` as a dependency.
-    let jitter = Duration::from_millis((now_ms & 0xFF) as u64);
+    let jitter = Duration::from_millis(now_ms & 0xFF);
 
     if let Some(reset_str) = reset_header {
         if let Ok(reset_ms) = reset_str.trim().parse::<u64>() {
@@ -1006,8 +1006,7 @@ pub fn openai_compat_request_body(req: &LlmRequest) -> serde_json::Value {
 /// retry, bubble up the typed error, or surface the response.
 enum OpenAiAttempt {
     Ok(LlmResponse),
-    /// 429. The dispatcher should honour `reset_at_ms`/`retry_after` when
-    /// scheduling the next attempt.
+    /// Rate-limited (429). The dispatcher should honour `reset_at_ms`/`retry_after` when scheduling the next attempt.
     RateLimited {
         status: u16,
         url: String,
